MAPeD: changed project file format (v6): tiles 4x4 data UINT -> ULONG; Updated Python API v1.8

+++ src/MAPeD/src/tiles_processor/tile_editor.cs
@@ -113,7 +113,7 @@ namespace MAPeD
 			{
 				// draw images...
 				{
-					byte tile_block;
+					ushort tile_block;
 					
 					for( int i = 0; i < utils.CONST_TILE_SIZE; i++ )
 					{
